Объединил с еще одним слайсом по изменению корзины. Удалиль лишьние запросы к API
