+++ src/entities/CartEntity/model/hooks/sliceHooks.ts
@@ -3,6 +3,3 @@ import { useSelector } from 'react-redux'
 import { StateSchema } from '@/app/providers/StoreProvider'
 
 export const useCartSelector = () => useSelector((store: StateSchema) => store.cartEntity)
-export const getCartSelector = (state: StateSchema) => {
-  return state.cartEntity.cart
-}
