+++ src/entities/CartEntity/model/selectors/selectors.ts
@@ -1,0 +1,5 @@
+import { StateSchema } from '@/app/providers/StoreProvider'
+
+export const getCartSelector = (state: StateSchema) => {
+  return state.cartEntity.cart
+}
