+++ src/entities/CartEntity/model/services/getCart.ts
@@ -3,20 +3,15 @@ import { createAsyncThunk } from '@reduxjs/toolkit'
 import { ThunkConfig } from '@/app/providers/StoreProvider/config/StateSchema'
 import { apiErrorIdentify } from '@/shared/api/apiErrorIdentify'
 import { ApiError, ApiErrorTypes, ApiRoutes } from '@/shared/api/types'
-import { IProductCartList } from '@/shared/model/types/ProductCartListModel'
 
-export const putIncreaseProductAmount = createAsyncThunk<IProductCartList, number, ThunkConfig<ApiError>>(
-  'cart-increase-product-amount',
-  async (productId, thunkAPI) => {
+import type { ICartEntity } from '../types/types'
+
+export const getCart = createAsyncThunk<ICartEntity, void, ThunkConfig<ApiError>>(
+  'cartEntitie/getCart',
+  async (_, thunkAPI) => {
     const { rejectWithValue, extra } = thunkAPI
     try {
-      const { data } = await extra.api.put(
-        `api/${ApiRoutes.INCREASE_PRODUCT_AMOUNT}`,
-        { product: productId },
-        {
-          withCredentials: true
-        }
-      )
+      const { data } = await extra.api.get(`api/${ApiRoutes.CART_LIST}/`)
       return data
     } catch (error) {
       return rejectWithValue(apiErrorIdentify(error, ApiErrorTypes.DATA_EMPTY_ERROR))
