+++ src/entities/CartEntity/model/slice/cartEntitySlice.ts
@@ -1,38 +1,14 @@
-import { createAsyncThunk, createSlice } from '@reduxjs/toolkit'
+import { createSlice } from '@reduxjs/toolkit'
 
-import type { ThunkConfig } from '@/app/providers/StoreProvider/config/StateSchema'
-import { apiErrorIdentify } from '@/shared/api/apiErrorIdentify'
 import { rejectedPayloadHandle } from '@/shared/api/rejectedPayloadHandle'
-import { ApiError, ApiErrorTypes, ApiRoutes } from '@/shared/api/types'
 
-import type { IAddedProduct, ICartEntity, ICartEntitySchema } from '../types/types'
-
-export const getCart = createAsyncThunk<ICartEntity, void, ThunkConfig<ApiError>>(
-  'cart/getCart',
-  async (_, thunkAPI) => {
-    const { rejectWithValue, extra } = thunkAPI
-    try {
-      const { data } = await extra.api.get(`api/${ApiRoutes.CART_LIST}/`)
-      return data
-    } catch (error) {
-      return rejectWithValue(apiErrorIdentify(error, ApiErrorTypes.DATA_EMPTY_ERROR))
-    }
-  }
-)
-
-export const addToCart = createAsyncThunk<void, IAddedProduct, ThunkConfig<ApiError>>(
-  'cart/addToCart',
-  async (addedProduct, thunkAPI) => {
-    const { rejectWithValue, extra } = thunkAPI
-    try {
-      await extra.api.post(`api/${ApiRoutes.CART_LIST}/`, addedProduct)
-
-      thunkAPI.dispatch(getCart())
-    } catch (error) {
-      return rejectWithValue(apiErrorIdentify(error, ApiErrorTypes.DATA_EMPTY_ERROR))
-    }
-  }
-)
+import { addToCart } from '../services/addToCart'
+import { getCart } from '../services/getCart'
+import { putDecreaseProductAmount } from '../services/putDecreaseProductAmount'
+import { putIncreaseProductAmount } from '../services/putIncreaseProductAmount'
+import { putRemoveProduct } from '../services/putRemoveProduct'
+import { putRenewProductAmount } from '../services/putRenewProductAmount'
+import type { ICartEntitySchema } from '../types/types'
 
 const initialState: ICartEntitySchema = {
   isLoading: false,
@@ -47,13 +23,14 @@ const initialState: ICartEntitySchema = {
 }
 
 export const cartEntitySlice = createSlice({
-  name: 'cart',
+  name: 'cartEntitie',
   initialState,
   reducers: {},
   extraReducers: builder => {
     builder
       .addCase(getCart.pending, state => {
         state.isLoading = true
+        state.error = null
       })
       .addCase(getCart.fulfilled, (state, { payload }) => {
         state.isLoading = false
@@ -66,11 +43,60 @@ export const cartEntitySlice = createSlice({
       builder
         .addCase(addToCart.pending, state => {
           state.isLoading = true
+          state.error = null
         })
         .addCase(addToCart.fulfilled, state => {
           state.isLoading = false
         })
         .addCase(addToCart.rejected, (state, { payload }) => {
+          state.isLoading = false
+          state.error = rejectedPayloadHandle(payload)
+        }),
+      builder
+        .addCase(putIncreaseProductAmount.pending, state => {
+          state.isLoading = true
+          state.error = null
+        })
+        .addCase(putIncreaseProductAmount.fulfilled, state => {
+          state.isLoading = false
+        })
+        .addCase(putIncreaseProductAmount.rejected, (state, { payload }) => {
+          state.isLoading = false
+          state.error = rejectedPayloadHandle(payload)
+        }),
+      builder
+        .addCase(putDecreaseProductAmount.pending, state => {
+          state.isLoading = true
+          state.error = null
+        })
+        .addCase(putDecreaseProductAmount.fulfilled, state => {
+          state.isLoading = false
+        })
+        .addCase(putDecreaseProductAmount.rejected, (state, { payload }) => {
+          state.isLoading = false
+          state.error = rejectedPayloadHandle(payload)
+        }),
+      builder
+        .addCase(putRenewProductAmount.pending, state => {
+          state.isLoading = true
+          state.error = null
+        })
+        .addCase(putRenewProductAmount.fulfilled, state => {
+          state.isLoading = false
+        })
+        .addCase(putRenewProductAmount.rejected, (state, { payload }) => {
+          state.isLoading = false
+          state.error = rejectedPayloadHandle(payload)
+        }),
+      builder
+        .addCase(putRemoveProduct.pending, state => {
+          state.isLoading = true
+          state.error = null
+        })
+        .addCase(putRemoveProduct.fulfilled, state => {
+          state.isLoading = false
+        })
+        .addCase(putRemoveProduct.rejected, (state, { payload }) => {
           state.isLoading = false
           state.error = rejectedPayloadHandle(payload)
         })
