--- src/features/CartEdit/model/selectors.ts
@@ -1,14 +1,0 @@
-import { StateSchema } from '@/app/providers/StoreProvider'
-
-export const isSuccessfulRequestSelector = (state: StateSchema) => {
-  return (
-    state.productAmount.isIncreaseSuccessful ||
-    state.productAmount.isDecreaseSuccessful ||
-    state.productAmount.isRenewProductAmountSuccessful ||
-    state.productAmount.isRemoveSuccessful
-  )
-}
-
-export const getProductListSelector = (state: StateSchema) => {
-  return state.productAmount.productList
-}
