+++ src/features/login/model/services/logout/logout.ts
@@ -1,6 +1,7 @@
 import { createAsyncThunk } from '@reduxjs/toolkit'
 
 import { ThunkConfig } from '@/app/providers/StoreProvider/config/StateSchema'
+import { getCart } from '@/entities/CartEntity/model/services/getCart'
 import { apiErrorIdentify } from '@/shared/api/apiErrorIdentify'
 import { ApiError, ApiErrorTypes, ApiRoutes } from '@/shared/api/types'
 import { LOCAL_STORAGE_TOKEN_KEY } from '@/shared/constants/localStorage'
@@ -17,6 +18,7 @@ export const logout = createAsyncThunk<void, void, ThunkConfig<ApiError>>(
       extra.api.delToken && extra.api.delToken()
       $localStorageHandler.removeItem(LOCAL_STORAGE_TOKEN_KEY)
 
+      thunkAPI.dispatch(getCart())
       dispatch(loginActions.userReset())
     } catch (error) {
       if (error) {
